# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para el desafío de nivel 3 de Fuego de Quasar.",
        "title": "Fuego de Quasar API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/topsecret": {
            "post": {
                "description": "Recibe información de los satélites y retorna posición y mensaje",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topsecret"
                ],
                "summary": "Decodifica mensaje y posición",
                "parameters": [
                    {
                        "description": "Datos de los satélites",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TopSecretRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ejemplo de respuesta\" example({\"position\":{\"x\":426.4001,\"y\":-252.80016},\"message\":\"este es un mensaje secreto\"})",
                        "schema": {
                            "$ref": "#/definitions/handlers.TopSecretResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/topsecret_split": {
            "get": {
                "description": "Recupera la posición y mensaje usando los datos guardados de los satélites",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topsecret_split"
                ],
                "summary": "Decodifica mensaje y posición usando información parcial",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.TopSecretResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/topsecret_split/{satellite_name}": {
            "post": {
                "description": "Permite guardar la distancia y mensaje de un satélite individualmente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "topsecret_split"
                ],
                "summary": "Guarda información parcial de un satélite",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nombre del satélite",
                        "name": "satellite_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Distancia y mensaje del satélite",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.TopSecretSplitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actualización exitosa"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.Position": {
            "description": "Coordenadas de la fuente",
            "type": "object",
            "properties": {
                "x": {
                    "type": "number",
                    "example": 426.4001
                },
                "y": {
                    "type": "number",
                    "example": -252.80016
                }
            }
        },
        "handlers.SatelliteInfo": {
            "description": "Información individual de un satélite",
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number",
                    "example": 927.75
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"este\"",
                        " \"\"",
                        " \"\"",
                        " \"mensaje\"",
                        " \"\"]"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "kenobi"
                }
            }
        },
        "handlers.TopSecretRequest": {
            "description": "Datos de los satélites para decodificar mensaje y posición",
            "type": "object",
            "properties": {
                "satellites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.SatelliteInfo"
                    }
                }
            }
        },
        "handlers.TopSecretResponse": {
            "description": "Respuesta con posición y mensaje decodificado",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "este es un mensaje secreto"
                },
                "position": {
                    "$ref": "#/definitions/handlers.Position"
                }
            }
        },
        "handlers.TopSecretSplitRequest": {
            "type": "object",
            "properties": {
                "distance": {
                    "type": "number"
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}